Credited test cases.
All that is left is to double check naming and comments

In the main method, I commented out the lines of code that ask for the command line args. Will need to make jar file to execute the simulation with arguments.

I coppied and pasted a lot so some comments might not make sense. Please feel free to change them